+++ src/main/scala/streams/tfc/package.scala
@@ -3,6 +3,7 @@ package streams
 import com.bioxx.tfc.Core.TFC_Climate._
 import com.bioxx.tfc.Core.TFC_Core._
 import com.bioxx.tfc.Core._
+import com.bioxx.tfc.Entities.EntityFallingBlockTFC
 import com.bioxx.tfc.WorldGen.Generators.WorldGenFissure
 import com.bioxx.tfc.WorldGen.TFCBiome._
 import farseek.block.BlockAndData
@@ -12,12 +13,17 @@ import java.util.Random
 import net.minecraft.block.Block
 import net.minecraft.world._
 import streams.block.BlockRiver
+import streams.world.gen.structure.RiverGenerator._
 
 package object tfc {
 
     def generate(generator: WorldGenFissure, world: World, random: Random, x: Int, y: Int, z: Int) {
         if(between(y, y+10).forall(!blockAt(x, _, z)(world).getMaterial.isLiquid))
             generator.generate(world, random, x, y, z)
+    }
+
+    def generateStreams(generator: com.bioxx.tfc.WorldGen.TFCChunkProviderGenerate, world: World, xChunk: Int, zChunk: Int, blocks: Array[Block], datas: Array[Byte]) {
+        surfaceWaterGenerator.onChunkGeneration(world.asInstanceOf[WorldServer], generator, xChunk, zChunk, blocks, datas)
     }
 
     def tfcSurfaceBlockAt(wxz: XZ, world: World, sedimentOnly: Boolean = false): BlockAndData = {
@@ -33,4 +39,16 @@ package object tfc {
 
     def isFreshWater(block: Block): Boolean = block.isInstanceOf[BlockRiver] || TFC_Core.isFreshWater(block)
 
+    def canReplace(fallingBlock: EntityFallingBlockTFC, world: World, x: Int, y: Int, z: Int): Boolean = {
+        if(!fallingBlock.isDead)
+            false
+        else {
+            if(!blockAt(x, y, z)(world).isInstanceOf[BlockRiver])
+                fallingBlock.canReplace(world, x, y, z)
+            else {
+                fallingBlock.shouldDropItem = true
+                false
+            }
+        }
+    }
 }