+++ src/main/scala/streams/world/gen/package.scala
@@ -1,13 +1,16 @@
 package streams.world
 
+import com.bioxx.tfc.WorldGen.TFCChunkProviderGenerate
 import farseek.util._
 import net.minecraft.world.WorldProvider
 import net.minecraft.world.chunk.IChunkProvider
-import streams.tfc.TFCChunkProviderGenerate
 
 package object gen {
 
+    private lazy val customTfc = classOf[TFCChunkProviderGenerate].getDeclaredMethods.exists(_.getName == "generateStreams")
+
     def createChunkGenerator(provider: WorldProvider): IChunkProvider =
-        if(tfcLoaded) new TFCChunkProviderGenerate(provider.worldObj, provider.worldObj.getSeed, provider.worldObj.getWorldInfo.isMapFeaturesEnabled)
+        if(tfcLoaded && provider.isSurfaceWorld && !customTfc)
+            new streams.tfc.TFCChunkProviderGenerate(provider.worldObj, provider.worldObj.getSeed, provider.worldObj.getWorldInfo.isMapFeaturesEnabled)
         else provider.createChunkGenerator
 }